feat: detail 모달 구현 및 기존 page render 코드 분리
